64: Languages 8 As a user I want to list the languages I know as well as the languages I want the person(s) I connect with to know.  The list should be sorted alphabetically with English at the top of the list.
NOTE TO BRIAN: There must be a list of common languages that can be used as a dropdown. However, I don't want every language on the planet, just the common ones. Use your best judgement and let me know what you find

65: Language Skill Level 4
As a user, for each language I specify for either me or the other person(s), I want to indicate the level of language skill: beginner, intermediate, or fluent